--- src/main/java/chess/dao/MyDataSource.java
@@ -1,10 +1,0 @@
-package chess.dao;
-
-import java.sql.Connection;
-
-public interface MyDataSource {
-
-    Connection getConnection();
-
-    void close();
-}
