--- src/main/java/chess/dao/TurnDao.java
@@ -1,9 +1,0 @@
-package chess.dao;
-
-
-public interface TurnDao {
-
-    String getCurrentTurn();
-
-    void updateTurn(final String currentTurn, final String previousTurn);
-}
